Rename TForce `RatesItemOptions`
This renames `RatesItemOptions` to `ItemOptions` for consistency
with the other services. The renamed class is aliased for backwards
compatibility, but also marked as deprecated.
