+++ lib/friendly_shipping/services/tforce_freight.rb
@@ -8,6 +8,7 @@ require 'friendly_shipping/services/tforce_freight/shipment_document'
 require 'friendly_shipping/services/tforce_freight/shipment_information'
 require 'friendly_shipping/services/tforce_freight/rates_options'
 require 'friendly_shipping/services/tforce_freight/rates_package_options'
+require 'friendly_shipping/services/tforce_freight/item_options'
 require 'friendly_shipping/services/tforce_freight/rates_item_options'
 require 'friendly_shipping/services/tforce_freight/pickup_options'
 require 'friendly_shipping/services/tforce_freight/bol_options'
