+++ lib/friendly_shipping/services/tforce_freight/rates_package_options.rb
@@ -38,7 +38,7 @@ module FriendlyShipping
           @handling_unit_code = HANDLING_UNIT_TYPES.fetch(handling_unit).fetch(:code)
           @handling_unit_description = HANDLING_UNIT_TYPES.fetch(handling_unit).fetch(:description)
           @handling_unit_tag = "handlingUnit#{HANDLING_UNIT_TYPES.fetch(handling_unit).fetch(:handling_unit_tag)}"
-          super(**kwargs.reverse_merge(item_options_class: RatesItemOptions))
+          super(**kwargs.reverse_merge(item_options_class: ItemOptions))
         end
       end
     end
