+++ spec/friendly_shipping/services/tforce_freight/generate_commodity_information_spec.rb
@@ -58,7 +58,7 @@ RSpec.describe FriendlyShipping::Services::TForceFreight::GenerateCommodityInfor
           FriendlyShipping::Services::TForceFreight::RatesPackageOptions.new(
             package_id: package.id,
             item_options: package.items.map do |item|
-              FriendlyShipping::Services::TForceFreight::RatesItemOptions.new(
+              FriendlyShipping::Services::TForceFreight::ItemOptions.new(
                 item_id: item.id,
                 packaging: :package,
                 freight_class: "92.5",
