+++ spec/friendly_shipping/services/tforce_freight/generate_create_bol_request_hash_spec.rb
@@ -101,7 +101,7 @@ RSpec.describe FriendlyShipping::Services::TForceFreight::GenerateCreateBOLReque
 
     let(:item_one_options) do
       [
-        FriendlyShipping::Services::TForceFreight::RatesItemOptions.new(
+        FriendlyShipping::Services::TForceFreight::ItemOptions.new(
           item_id: "item_one",
           packaging: :pallet,
           freight_class: "92.5",
